Fix glDeleteSync validation
glDeleteSync can only delete an existing FenceSync object.

+++ src/OpenGL/libGLESv2/libGLESv3.cpp
@@ -3004,6 +3004,11 @@ GL_APICALL void GL_APIENTRY glDeleteSync(GLsync sync)
 
 	if(context)
 	{
+		if(sync && !context->getFenceSync(sync))
+		{
+			return error(GL_INVALID_VALUE);
+		}
+
 		context->deleteFenceSync(sync);
 	}
 }
